Update main.m
integer cast when calculating the expected final amount

+++ code/main.m
@@ -22,7 +22,7 @@ M=montecarlo_matrix(initial_amount,bet_amount, n_bet,n_rep);
 plot_montecarlo(M);
 %--------------------------------------------------------------------------
 % EXPECTED FINAL AMOUNT
-expected_final_amount= mean(M(end,:));
+expected_final_amount= int64(mean(M(end,:)));
 %--------------------------------------------------------------------------
 % EXPECTED LOSS
 expected_loss= initial_amount - expected_final_amount;
